Add timeout in case game does not terminate
This patch adds a timeout around the main game loop. For certain inputs,
it is possible there is a cycle in which no operations progress the
game. For instance, this happens when no operation is applicable, or
when bots keep passing around values without them ever ending up at the
outputs. This timeout protects us from that.