Provide executable to answer challenge questions
This patch adds a new executable that outputs answers to some questions
in a challenge. The challenge questions are added to the README.

To make finding a bot that compares two chip values easier, a new log
item was introduced: the bot compare log items. This log item contains
the ID of the bot and the values compared, and is added just before the
bot gives the chips to other entities.