Make game state format output whole relevant world
This patch updates the format method on the game state to output all
entities to that have any values. While it does not have any impact if
the game has ended, it does make the format method more useful to
inspect the game state at a certain point in time when the game is still
running.