Group classes into modules
This patch groups a couple of classes into modules when they represent
similar things. As of now, all classes within a submodule all have the
same interface. This makes it easier to see all classes that belong
together and makes navigating to the correct file easier.

The following classes were grouped:

* Operations were renamed to instructions and grouped under
  instructions.
* Log items were grouped under log.
* Entities were grouped under entities.