Split instructions by type in parser
This patch makes the parser return two sets of results in case of
success: a set of input instructions and a set of bot instructions. This
change makes the pre validator and game a lot simpler: they both used to
split the instructions by type, but this is no longer necessary since
they received the split sets.